fix(tui): use int64 for statistics counters to prevent overflow
- Change TotalPackets from int to int64 in Statistics struct
- Change TotalPackets and MatchedPackets from int to int64 in PacketStore
- Change BoundedCounter.counts from map[string]int to map[string]int64
- Update all usages to use int64 casts where needed
- Update tests to use int64 assertions

This ensures consistent behavior across 32-bit and 64-bit platforms
and prevents potential overflow for long-running capture sessions.

Addresses issue 16 from code review fixes plan.

+++ internal/pkg/tui/filter_helpers.go
@@ -147,7 +147,7 @@ func isBPFExpression(s string) bool {
 // applyFilters re-applies all active filters to the packet list
 func (m *Model) applyFilters() {
 	if !m.packetStore.HasFilter() {
-		m.packetStore.MatchedPackets = m.packetStore.PacketsCount
+		m.packetStore.MatchedPackets = int64(m.packetStore.PacketsCount)
 		m.packetStore.FilteredPackets = make([]components.PacketDisplay, 0)
 		return
 	}
@@ -159,5 +159,5 @@ func (m *Model) applyFilters() {
 			m.packetStore.FilteredPackets = append(m.packetStore.FilteredPackets, pkt)
 		}
 	}
-	m.packetStore.MatchedPackets = len(m.packetStore.FilteredPackets)
+	m.packetStore.MatchedPackets = int64(len(m.packetStore.FilteredPackets))
 }
